chore: clean up and rename of a few components and files
